requirements.txt

pymodbus>=2.5.0
pyserial>=3.4
